requirements.txt
flask
openai
requests
configparser
langchain[all]
gunicorn
pinecone-client
tiktoken
uwsgi
pandas